Improve client for discoverability
- Move general ontology and data series discovery features from sample module  up to main client module,
- Make separate crop model module for crop weighting functions

The changes are backward compatible with existing client code.
